+++ src/client/customers.rs
@@ -17,9 +17,10 @@ use codes_iso_3166::part_1::CountryCode;
 use codes_iso_4217::CurrencyCode;
 use futures_core::Stream;
 use futures_util::stream::TryStreamExt;
-use reqwest::Method;
+use reqwest::{Method, RequestBuilder};
 use serde::{Deserialize, Serialize};
 use serde_enum_str::{Deserialize_enum_str, Serialize_enum_str};
+use time::format_description::well_known::Rfc3339;
 use time::OffsetDateTime;
 
 use crate::client::taxes::{TaxId, TaxIdRequest};
@@ -30,6 +31,11 @@ use crate::serde::Empty;
 use crate::util::StrIteratorExt;
 
 const CUSTOMERS_PATH: [&str; 1] = ["customers"];
+
+#[derive(Deserialize)]
+struct ArrayResponse<T> {
+    data: Vec<T>,
+}
 
 /// A customer ID.
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize)]
@@ -309,6 +315,7 @@ pub struct AddVoidCreditLedgerEntryRequestParams<'a> {
     pub description: Option<&'a str>,
 }
 
+/// A block of credit held by a customer.
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Deserialize, Serialize)]
 pub struct CustomerCreditBlock {
     /// The Orb-assigned unique identifier for the credit block.
@@ -430,6 +437,208 @@ pub struct VoidInitiatedLedgerEntry {
     pub void_reason: Option<String>,
     /// The amount voided from the ledger.
     pub void_amount: serde_json::Number,
+}
+
+/// The view mode for a cost breakdown.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Deserialize_enum_str, Serialize_enum_str)]
+pub enum CostViewMode {
+    /// Provide results as an incremental day-by-day view.
+    #[serde(rename = "periodic")]
+    Periodic,
+    /// Provide results as cumulative totals since the start of the billing period.
+    #[serde(rename = "cumulative")]
+    Cumulative,
+}
+
+#[derive(Debug, Default, Clone)]
+struct CustomerCostParamsFilter<'a> {
+    timeframe_start: Option<&'a OffsetDateTime>,
+    timeframe_end: Option<&'a OffsetDateTime>,
+    view_mode: Option<CostViewMode>,
+    group_by: Option<&'a str>,
+}
+
+trait Filterable<T> {
+    /// Apply the filter to a request.
+    fn apply(self, filter: &T) -> Self;
+}
+
+impl Filterable<CustomerCostParamsFilter<'_>> for RequestBuilder {
+    /// Apply the filter to a request.
+    fn apply(mut self, filter: &CustomerCostParamsFilter) -> Self {
+        if let Some(view_mode) = &filter.view_mode {
+            self = self.query(&[("view_mode", view_mode.to_string())]);
+        }
+        if let Some(group_by) = &filter.group_by {
+            self = self.query(&[("group_by", group_by)]);
+        }
+        if let Some(timeframe_start) = &filter.timeframe_start {
+            self = self.query(&[("timeframe_start", timeframe_start.format(&Rfc3339).unwrap())]);
+        }
+        if let Some(timeframe_end) = &filter.timeframe_end {
+            self = self.query(&[("timeframe_start", timeframe_end.format(&Rfc3339).unwrap())]);
+        }
+        self
+    }
+}
+
+/// Parameters for a Customer Costs query.
+#[derive(Debug, Default, Clone)]
+pub struct CustomerCostParams<'a> {
+    filter: CustomerCostParamsFilter<'a>,
+}
+
+impl<'a> CustomerCostParams<'a> {
+    /// The start of the returned range. If not specified this defaults to the billing period start
+    /// date.
+    pub const fn timeframe_start(mut self, timeframe_start: &'a OffsetDateTime) -> Self {
+        self.filter.timeframe_start = Some(timeframe_start);
+        self
+    }
+
+    /// The end of the returned range. If unspecified will default to the billing period end date.
+    pub const fn timeframe_end(mut self, timeframe_end: &'a OffsetDateTime) -> Self {
+        self.filter.timeframe_end = Some(timeframe_end);
+        self
+    }
+
+    /// How costs should be broken down in the resultant day-by-day view.
+    pub const fn view_mode(mut self, view_mode: CostViewMode) -> Self {
+        self.filter.view_mode = Some(view_mode);
+        self
+    }
+
+    /// The custom attribute to group costs by.
+    pub const fn group_by(mut self, group_by: &'a str) -> Self {
+        self.filter.group_by = Some(group_by);
+        self
+    }
+}
+
+/// A group of costs for a given timeframe.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub struct CustomerCostBucket {
+    /// Total costs for the timeframe, excluding any minimums and discounts.
+    pub subtotal: String,
+    /// Total costs for the timeframe, including any minimums and discounts.
+    pub total: String,
+    /// The starting point for the timeframe.
+    #[serde(with = "time::serde::rfc3339")]
+    pub timeframe_start: OffsetDateTime,
+    /// The ending point for the timeframe.
+    #[serde(with = "time::serde::rfc3339")]
+    pub timeframe_end: OffsetDateTime,
+    /// The costs for each price.
+    pub per_price_costs: Vec<CustomerCostPriceBlock>,
+}
+
+/// The cost for a given Price within a timeframe.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub struct CustomerCostPriceBlock {
+    /// The price's quantity for the timeframe.
+    pub quantity: Option<serde_json::Number>,
+    /// The price's contributions for the timeframe, excluding any minimums and discounts.
+    pub subtotal: String,
+    /// The price's contributions for the timeframe, including any minimums and discounts.
+    pub total: String,
+    /// The price that can be billed on a subscription.
+    pub price: CustomerCostPriceBlockPrice,
+    /// The price costs per grouping key.
+    pub price_groups: Option<Vec<CustomerCostPriceBlockPriceGroup>>,
+}
+
+/// A price cost for a given set of grouping keys.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub struct CustomerCostPriceBlockPriceGroup {
+    /// The key breaking down a single price's costs.
+    pub grouping_key: String,
+    /// An optional value for the key.
+    pub grouping_value: Option<String>,
+    /// The second dimension for the matrix price, if applicable.
+    pub secondary_grouping_key: Option<String>,
+    /// An optional value for the `secondary_grouping_key`, if applicable.
+    pub secondary_grouping_value: Option<String>,
+    /// Total costs for this group for the timeframe, excluding any minimums and discounts.
+    // this should be thought of as a "subtotal" to align with the rest of the API, but we're
+    // keeping the existing Orb terminology.
+    pub total: String,
+}
+
+/// The type of pricing
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[serde(tag = "model_type")]
+pub enum CustomerCostPriceBlockPrice {
+    /// Sets of unit prices in a one or two-dimensional matrix.
+    #[serde(rename = "matrix")]
+    Matrix(CustomerCostPriceBlockMatrixPrice),
+    /// A fixed amount per unit of usage.
+    #[serde(rename = "unit")]
+    Unit(CustomerCostPriceBlockUnitPrice),
+    // TODO: Add support for additional pricing models
+}
+
+/// Matrix pricing details.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub struct CustomerCostPriceBlockMatrixPrice {
+    /// The Orb-assigned unique identifier for the price block.
+    pub id: String,
+    /// An optional user-defined ID for this price resource.
+    #[serde(rename = "external_price_id")]
+    pub external_id: Option<String>,
+    /// Information about the item being priced.
+    pub item: CustomerCostItem,
+    /// The configuration for this matrix price.
+    pub matrix_config: CustomerCostPriceBlockMatrixPriceConfig,
+}
+
+/// An item being priced.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub struct CustomerCostItem {
+    /// Orb's unique identifier for the item.
+    pub id: String,
+    /// The item's name.
+    pub name: String,
+}
+
+/// Configuration for a pricing matrix.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub struct CustomerCostPriceBlockMatrixPriceConfig {
+    /// The fallback unit amount.
+    pub default_unit_amount: String,
+    /// A collection of dimensions modeled by the matrix.
+    pub dimensions: Vec<String>,
+    /// All pricing values configured for the matrix.
+    pub matrix_values: Vec<CustomerCostPriceBlockMatrixPriceValue>,
+}
+
+/// A pricing value for a cell within the pricing matrix.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub struct CustomerCostPriceBlockMatrixPriceValue {
+    /// The dimensions corresponding to this cell.
+    pub dimension_values: Vec<String>,
+    /// The per-unit amount usage within this cell bills.
+    pub unit_amount: String,
+}
+
+/// Unit pricing details.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub struct CustomerCostPriceBlockUnitPrice {
+    /// The Orb-assigned unique identifier for the price block.
+    pub id: String,
+    /// An optional user-defined ID for this price resource.
+    #[serde(rename = "external_price_id")]
+    pub external_id: Option<String>,
+    /// Information about the item being priced.
+    pub item: CustomerCostItem,
+    /// The configuration for this unit price.
+    pub unit_config: CustomerCostPriceBlockUnitPriceConfig,
+}
+
+/// Configuration for a unit price.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub struct CustomerCostPriceBlockUnitPriceConfig {
+    /// Per-unit pricing.
+    pub unit_amount: String,
 }
 
 impl Client {
@@ -566,7 +775,7 @@ impl Client {
         self.stream_paginated_request(params, req)
     }
 
-    /// Ceate a new ledger entry for the specified customer's balance.
+    /// Create a new ledger entry for the specified customer's balance.
     pub async fn create_ledger_entry(
         &self,
         id: &str,
@@ -580,7 +789,36 @@ impl Client {
                 .chain_one("ledger_entry"),
         );
         let req = req.json(entry);
-        let res = self.send_request(req).await?;
-        Ok(res)
+        self.send_request(req).await
+    }
+
+    /// Fetch a day-by-day snapshot of a customer's costs.
+    pub async fn get_customer_costs(
+        &self,
+        id: &str,
+        params: &CustomerCostParams<'_>,
+    ) -> Result<Vec<CustomerCostBucket>, Error> {
+        let req = self.build_request(Method::GET, CUSTOMERS_PATH.chain_one(id).chain_one("costs"));
+        let req = req.apply(&params.filter);
+        let res: ArrayResponse<CustomerCostBucket> = self.send_request(req).await?;
+        Ok(res.data)
+    }
+
+    /// Fetch a day-by-day snapshot of a customer's costs by their external ID.
+    pub async fn get_customer_costs_by_external_id(
+        &self,
+        external_id: &str,
+        params: &CustomerCostParams<'_>,
+    ) -> Result<Vec<CustomerCostBucket>, Error> {
+        let req = self.build_request(
+            Method::GET,
+            CUSTOMERS_PATH
+                .chain_one("external_customer_id")
+                .chain_one(external_id)
+                .chain_one("costs"),
+        );
+        let req = req.apply(&params.filter);
+        let res: ArrayResponse<CustomerCostBucket> = self.send_request(req).await?;
+        Ok(res.data)
     }
 }